% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_write.R
\name{.project_template_with_offset}
\alias{.project_template_with_offset}
\title{Project data onto a shared latent template and recover any scan offset}
\usage{
.project_template_with_offset(template, data, template_loadings)
}
\description{
Project data onto a shared latent template and recover any scan offset
}
\keyword{internal}
